# tool/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/add-single-person-and-match": {
            "post": {
                "description": "加入用戶和配對",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tinder"
                ],
                "summary": "加入用戶",
                "parameters": [
                    {
                        "description": "raw",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddSinglePersonAndMatchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    }
                }
            }
        },
        "/v1/query-single-people": {
            "get": {
                "description": "配對n組",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tinder"
                ],
                "summary": "配對n組",
                "parameters": [
                    {
                        "description": "raw",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.QuerySinglePeopleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    }
                }
            }
        },
        "/v1/remove-single-person": {
            "post": {
                "description": "移除用戶",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tinder"
                ],
                "summary": "移除用戶",
                "parameters": [
                    {
                        "description": "raw",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RemoveSinglePersonReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "$ref": "#/definitions/app.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.BaseResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "result": {}
            }
        },
        "dto.AddSinglePersonAndMatchReq": {
            "type": "object",
            "required": [
                "date_counts",
                "gender",
                "height",
                "name"
            ],
            "properties": {
                "date_counts": {
                    "type": "integer"
                },
                "gender": {
                    "type": "integer"
                },
                "height": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.QuerySinglePeopleReq": {
            "type": "object",
            "properties": {
                "counts": {
                    "type": "integer"
                }
            }
        },
        "dto.RemoveSinglePersonReq": {
            "type": "object",
            "required": [
                "gender",
                "name"
            ],
            "properties": {
                "gender": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}